Update Profile page prepared for backend work on profile data
